+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinExternalContact/CorpTag/CgibinExternalContactAddCorpTagRequest.cs
@@ -24,7 +24,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("order")]
                 [System.Text.Json.Serialization.JsonPropertyName("order")]
-                public int? Order { get; set; }
+                public long? Order { get; set; }
             }
         }
 
@@ -47,7 +47,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("order")]
         [System.Text.Json.Serialization.JsonPropertyName("order")]
-        public int? Order { get; set; }
+        public long? Order { get; set; }
 
         /// <summary>
         /// 获取或设置企业标签列表。
