+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinExternalContact/CorpTag/CgibinExternalContactEditCorpTagRequest.cs
@@ -27,7 +27,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("order")]
         [System.Text.Json.Serialization.JsonPropertyName("order")]
-        public int? Order { get; set; }
+        public long? Order { get; set; }
 
         /// <summary>
         /// 获取或设置授权方安装的应用 ID。
