+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinExternalContact/CorpTag/CgibinExternalContactGetCorpTagListResponse.cs
@@ -35,7 +35,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("order")]
                         [System.Text.Json.Serialization.JsonPropertyName("order")]
-                        public int Order { get; set; }
+                        public long Order { get; set; }
 
                         /// <summary>
                         /// 获取或设置是否已经被删除。
@@ -72,7 +72,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("order")]
                 [System.Text.Json.Serialization.JsonPropertyName("order")]
-                public int Order { get; set; }
+                public long Order { get; set; }
 
                 /// <summary>
                 /// 获取或设置企业标签列表。
