+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinLinkedCorp/CgibinLinkedCorpDepartmentListResponse.cs
@@ -38,7 +38,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("order")]
                 [System.Text.Json.Serialization.JsonPropertyName("order")]
-                public int DepartmentOrder { get; set; }
+                public long DepartmentOrder { get; set; }
             }
         }
